requirements.txt
crewai
langchain-openai
python-dotenv
requests
beautifulsoup4
duckduckgo-search
